refactor: define the own server-side constants
