+++ src/argilla/server/constants.py
@@ -1,0 +1,18 @@
+
+API_KEY_HEADER_NAME = "X-Argilla-Api-Key"
+WORKSPACE_HEADER_NAME = "X-Argilla-Workspace"
+
+DEFAULT_USERNAME = "argilla"
+DEFAULT_PASSWORD = "1234"
+DEFAULT_API_KEY = "argilla.apikey"
+
+DEFAULT_MAX_KEYWORD_LENGTH = 128
+DEFAULT_TELEMETRY_KEY = "C6FkcaoCbt78rACAgvyBxGBcMB3dM3nn"
+
+
+# The metadata field name prefix defined for protected (non-searchable) values
+PROTECTED_METADATA_FIELD_PREFIX = "_"
+
+ES_INDEX_REGEX_PATTERN = r"^(?!-|_)[a-z0-9-_]+$"
+
+_JS_MAX_SAFE_INTEGER = 9007199254740991
