+++ src/argilla/server/daos/backend/generic_elastic.py
@@ -14,7 +14,7 @@
 
 from typing import Any, Dict, Iterable, List, Optional, Tuple
 
-from argilla._constants import PROTECTED_METADATA_FIELD_PREFIX
+from argilla.server.constants import PROTECTED_METADATA_FIELD_PREFIX
 from argilla.logging import LoggingMixin
 from argilla.server.commons.models import TaskType
 from argilla.server.daos.backend.base import IndexNotFoundError, InvalidSearchError
