+++ src/argilla/server/daos/models/records.py
@@ -19,7 +19,7 @@ from datetime import datetime
 from typing import Any, Dict, Generic, List, Optional, TypeVar, Union
 
 from argilla import _messages
-from argilla._constants import _JS_MAX_SAFE_INTEGER, PROTECTED_METADATA_FIELD_PREFIX
+from argilla.server.constants import _JS_MAX_SAFE_INTEGER, PROTECTED_METADATA_FIELD_PREFIX
 from argilla.server.commons.models import PredictionStatus, TaskStatus, TaskType
 from argilla.server.daos.backend.search.model import BaseRecordsQuery, SortConfig
 from argilla.server.helpers import flatten_dict
