+++ src/argilla/server/schemas/v0/datasets.py
@@ -20,7 +20,7 @@ from datetime import datetime
 from typing import Any, Dict, Optional, Union
 from uuid import UUID
 
-from argilla._constants import ES_INDEX_REGEX_PATTERN
+from argilla.server.constants import ES_INDEX_REGEX_PATTERN
 from argilla.server.commons.models import TaskType
 from argilla.server.pydantic_v1 import BaseModel, Field
