+++ src/argilla/server/security/auth_provider/base.py
@@ -18,7 +18,7 @@ from typing import Optional
 from fastapi import Depends, FastAPI, Request
 from fastapi.security import APIKeyHeader, SecurityScopes
 
-from argilla._constants import API_KEY_HEADER_NAME
+from argilla.server.constants import API_KEY_HEADER_NAME
 from argilla.server.models import User
 
 api_key_header = APIKeyHeader(name=API_KEY_HEADER_NAME, auto_error=False)
