+++ src/argilla/server/settings.py
@@ -24,7 +24,7 @@ from pathlib import Path
 from typing import List, Optional
 from urllib.parse import urlparse
 
-from argilla._constants import DEFAULT_MAX_KEYWORD_LENGTH, DEFAULT_TELEMETRY_KEY
+from argilla.server.constants import DEFAULT_MAX_KEYWORD_LENGTH, DEFAULT_TELEMETRY_KEY
 from argilla.server.pydantic_v1 import BaseSettings, Field, root_validator, validator
